+++ Agda.cabal
@@ -360,6 +360,9 @@ library
                     Agda.Syntax.Common
                     Agda.Syntax.Concrete.Attribute
                     Agda.Syntax.Concrete.Definitions
+                    Agda.Syntax.Concrete.Definitions.Errors
+                    Agda.Syntax.Concrete.Definitions.Monad
+                    Agda.Syntax.Concrete.Definitions.Types
                     Agda.Syntax.Concrete.Fixity
                     Agda.Syntax.Concrete.Generic
                     Agda.Syntax.Concrete.Glyph
