+++ src/full/Agda/Syntax/Concrete/Definitions.hs
@@ -21,6 +21,16 @@
 --     A block starts when a lone signature is encountered, and ends when
 --     all lone signatures have seen their definition.
 --
+--   * Handle infix mutual blocks.
+--     In an `infix mutual' block we:
+--     * leave the data and fun sigs in place
+--     * classify signatures in `constructor' block based on their return type
+--       and group them all as a data def at the position in the block where the
+--       first constructor for the data sig in question occured
+--     * classify fun clauses based on the declared function used and group them
+--       all as a fundef at the position in the block where the first such fun
+--       clause appeared
+--
 --   * Report basic well-formedness error,
 --     when one of the above transformation fails.
 --     When possible, errors should be deferred to the scope checking phase
@@ -65,8 +75,13 @@ import Agda.Syntax.Common hiding (TerminationCheck())
 import qualified Agda.Syntax.Common as Common
 import Agda.Syntax.Position
 import Agda.Syntax.Notation
+import Agda.Syntax.Concrete.Definitions.Errors
 import Agda.Syntax.Concrete.Pretty () --instance only
 import Agda.Syntax.Concrete.Fixity
+
+import Agda.Syntax.Concrete.Definitions.Errors
+import Agda.Syntax.Concrete.Definitions.Monad
+import Agda.Syntax.Concrete.Definitions.Types
 
 import Agda.Interaction.Options.Warnings
 
@@ -88,559 +103,8 @@ import Agda.Utils.Update
 import Agda.Utils.Impossible
 
 {--------------------------------------------------------------------------
-    Types
- --------------------------------------------------------------------------}
-
-{-| The nice declarations. No fixity declarations and function definitions are
-    contained in a single constructor instead of spread out between type
-    signatures and clauses. The @private@, @postulate@, @abstract@ and @instance@
-    modifiers have been distributed to the individual declarations.
-
-    Observe the order of components:
-
-      Range
-      Fixity'
-      Access
-      IsAbstract
-      IsInstance
-      TerminationCheck
-      PositivityCheck
-
-      further attributes
-
-      (Q)Name
-
-      content (Expr, Declaration ...)
--}
-data NiceDeclaration
-  = Axiom Range Access IsAbstract IsInstance ArgInfo Name Expr
-      -- ^ 'IsAbstract' argument: We record whether a declaration was made in an @abstract@ block.
-      --
-      --   'ArgInfo' argument: Axioms and functions can be declared irrelevant.
-      --   ('Hiding' should be 'NotHidden'.)
-  | NiceField Range Access IsAbstract IsInstance TacticAttribute Name (Arg Expr)
-  | PrimitiveFunction Range Access IsAbstract Name (Arg Expr)
-  | NiceMutual Range TerminationCheck CoverageCheck PositivityCheck [NiceDeclaration]
-  | NiceModule Range Access IsAbstract QName Telescope [Declaration]
-  | NiceModuleMacro Range Access Name ModuleApplication OpenShortHand ImportDirective
-  | NiceOpen Range QName ImportDirective
-  | NiceImport Range QName (Maybe AsName) OpenShortHand ImportDirective
-  | NicePragma Range Pragma
-  | NiceRecSig Range Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr
-  | NiceDataSig Range Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr
-  | NiceFunClause Range Access IsAbstract TerminationCheck CoverageCheck Catchall Declaration
-    -- ^ An uncategorized function clause, could be a function clause
-    --   without type signature or a pattern lhs (e.g. for irrefutable let).
-    --   The 'Declaration' is the actual 'FunClause'.
-  | FunSig Range Access IsAbstract IsInstance IsMacro ArgInfo TerminationCheck CoverageCheck Name Expr
-  | FunDef Range [Declaration] IsAbstract IsInstance TerminationCheck CoverageCheck Name [Clause]
-      -- ^ Block of function clauses (we have seen the type signature before).
-      --   The 'Declaration's are the original declarations that were processed
-      --   into this 'FunDef' and are only used in 'notSoNiceDeclaration'.
-      --   Andreas, 2017-01-01: Because of issue #2372, we add 'IsInstance' here.
-      --   An alias should know that it is an instance.
-  | NiceDataDef Range Origin IsAbstract PositivityCheck UniverseCheck Name [LamBinding] [NiceConstructor]
-  | NiceLoneConstructor Range [NiceConstructor]
-  | NiceRecDef Range Origin IsAbstract PositivityCheck UniverseCheck Name RecordDirectives [LamBinding] [Declaration]
-      -- ^ @(Maybe Range)@ gives range of the 'pattern' declaration.
-  | NicePatternSyn Range Access Name [Arg Name] Pattern
-  | NiceGeneralize Range Access ArgInfo TacticAttribute Name Expr
-  | NiceUnquoteDecl Range Access IsAbstract IsInstance TerminationCheck CoverageCheck [Name] Expr
-  | NiceUnquoteDef Range Access IsAbstract TerminationCheck CoverageCheck [Name] Expr
-  deriving (Data, Show)
-
-type TerminationCheck = Common.TerminationCheck Measure
-
--- | Termination measure is, for now, a variable name.
-type Measure = Name
-
-type Catchall = Bool
-
--- | Only 'Axiom's.
-type NiceConstructor = NiceTypeSignature
-
--- | Only 'Axiom's.
-type NiceTypeSignature  = NiceDeclaration
-
--- | One clause in a function definition. There is no guarantee that the 'LHS'
---   actually declares the 'Name'. We will have to check that later.
-data Clause = Clause Name Catchall LHS RHS WhereClause [Clause]
-    deriving (Data, Show)
-
--- | Exception with internal source code callstack
-data DeclarationException = DeclarationException
-  { deLocation  :: CallStack
-  , deException :: DeclarationException'
-  }
-
-declarationException :: HasCallStack => DeclarationException' -> Nice a
-declarationException e = withCallerCallStack $ throwError . flip DeclarationException e
-
--- | The exception type.
-data DeclarationException'
-  = MultipleEllipses Pattern
-  | InvalidName Name
-  | DuplicateDefinition Name
-  | DuplicateAnonDeclaration Range
-  | MissingWithClauses Name LHS
-  | WrongDefinition Name DataRecOrFun DataRecOrFun
-  | DeclarationPanic String
-  | WrongContentBlock KindOfBlock Range
-  | AmbiguousFunClauses LHS (List1 Name)
-      -- ^ In a mutual block, a clause could belong to any of the ≥2 type signatures ('Name').
-  | InvalidMeasureMutual Range
-      -- ^ In a mutual block, all or none need a MEASURE pragma.
-      --   Range is of mutual block.
-  | UnquoteDefRequiresSignature (List1 Name)
-  | BadMacroDef NiceDeclaration
-    deriving (Data, Show)
-
-
-data DeclarationWarning = DeclarationWarning
-  { dwLocation :: CallStack
-  , dwWarning  :: DeclarationWarning'
-  } deriving (Show)
-
-declarationWarning' :: DeclarationWarning' -> CallStack -> Nice ()
-declarationWarning' w loc = niceWarning $ DeclarationWarning loc w
-
-declarationWarning :: HasCallStack => DeclarationWarning' -> Nice ()
-declarationWarning = withCallerCallStack . declarationWarning'
-
--- | Non-fatal errors encountered in the Nicifier.
-data DeclarationWarning'
-  -- Please keep in alphabetical order.
-  = EmptyAbstract Range    -- ^ Empty @abstract@  block.
-  | EmptyConstructor Range -- ^ Empty @constructor@ block.
-  | EmptyField Range       -- ^ Empty @field@     block.
-  | EmptyGeneralize Range  -- ^ Empty @variable@  block.
-  | EmptyInstance Range    -- ^ Empty @instance@  block
-  | EmptyMacro Range       -- ^ Empty @macro@     block.
-  | EmptyMutual Range      -- ^ Empty @mutual@    block.
-  | EmptyPostulate Range   -- ^ Empty @postulate@ block.
-  | EmptyPrivate Range     -- ^ Empty @private@   block.
-  | EmptyPrimitive Range   -- ^ Empty @primitive@ block.
-  | InvalidCatchallPragma Range
-      -- ^ A {-\# CATCHALL \#-} pragma
-      --   that does not precede a function clause.
-  | InvalidConstructor Range
-      -- ^ Invalid definition in a constructor block
-  | InvalidConstructorBlock Range
-      -- ^ Invalid constructor block (not inside an infix mutual block)
-  | InvalidCoverageCheckPragma Range
-      -- ^ A {-\# NON_COVERING \#-} pragma that does not apply to any function.
-  | InvalidNoPositivityCheckPragma Range
-      -- ^ A {-\# NO_POSITIVITY_CHECK \#-} pragma
-      --   that does not apply to any data or record type.
-  | InvalidNoUniverseCheckPragma Range
-      -- ^ A {-\# NO_UNIVERSE_CHECK \#-} pragma
-      --   that does not apply to a data or record type.
-  | InvalidRecordDirective Range
-      -- ^ A record directive outside of a record / below existing fields.
-  | InvalidTerminationCheckPragma Range
-      -- ^ A {-\# TERMINATING \#-} and {-\# NON_TERMINATING \#-} pragma
-      --   that does not apply to any function.
-  | MissingDefinitions [(Name, Range)]
-      -- ^ Declarations (e.g. type signatures) without a definition.
-  | NotAllowedInMutual Range String
-  | OpenPublicPrivate Range
-      -- ^ @private@ has no effect on @open public@.  (But the user might think so.)
-  | OpenPublicAbstract Range
-      -- ^ @abstract@ has no effect on @open public@.  (But the user might think so.)
-  | PolarityPragmasButNotPostulates [Name]
-  | PragmaNoTerminationCheck Range
-  -- ^ Pragma @{-\# NO_TERMINATION_CHECK \#-}@ has been replaced
-  --   by @{-\# TERMINATING \#-}@ and @{-\# NON_TERMINATING \#-}@.
-  | PragmaCompiled Range
-  -- ^ @COMPILE@ pragmas are not allowed in safe mode
-  | ShadowingInTelescope [(Name, [Range])]
-  | UnknownFixityInMixfixDecl [Name]
-  | UnknownNamesInFixityDecl [Name]
-  | UnknownNamesInPolarityPragmas [Name]
-  | UselessAbstract Range
-      -- ^ @abstract@ block with nothing that can (newly) be made abstract.
-  | UselessInstance Range
-      -- ^ @instance@ block with nothing that can (newly) become an instance.
-  | UselessPrivate Range
-      -- ^ @private@ block with nothing that can (newly) be made private.
-  deriving (Data, Show)
-
-declarationWarningName :: DeclarationWarning -> WarningName
-declarationWarningName = declarationWarningName' . dwWarning
-
-declarationWarningName' :: DeclarationWarning' -> WarningName
-declarationWarningName' = \case
-  -- Please keep in alphabetical order.
-  EmptyAbstract{}                   -> EmptyAbstract_
-  EmptyConstructor{}                -> EmptyConstructor_
-  EmptyField{}                      -> EmptyField_
-  EmptyGeneralize{}                 -> EmptyGeneralize_
-  EmptyInstance{}                   -> EmptyInstance_
-  EmptyMacro{}                      -> EmptyMacro_
-  EmptyMutual{}                     -> EmptyMutual_
-  EmptyPrivate{}                    -> EmptyPrivate_
-  EmptyPostulate{}                  -> EmptyPostulate_
-  EmptyPrimitive{}                  -> EmptyPrimitive_
-  InvalidCatchallPragma{}           -> InvalidCatchallPragma_
-  InvalidConstructor{}              -> InvalidConstructor_
-  InvalidConstructorBlock{}         -> InvalidConstructorBlock_
-  InvalidNoPositivityCheckPragma{}  -> InvalidNoPositivityCheckPragma_
-  InvalidNoUniverseCheckPragma{}    -> InvalidNoUniverseCheckPragma_
-  InvalidRecordDirective{}          -> InvalidRecordDirective_
-  InvalidTerminationCheckPragma{}   -> InvalidTerminationCheckPragma_
-  InvalidCoverageCheckPragma{}      -> InvalidCoverageCheckPragma_
-  MissingDefinitions{}              -> MissingDefinitions_
-  NotAllowedInMutual{}              -> NotAllowedInMutual_
-  OpenPublicPrivate{}               -> OpenPublicPrivate_
-  OpenPublicAbstract{}              -> OpenPublicAbstract_
-  PolarityPragmasButNotPostulates{} -> PolarityPragmasButNotPostulates_
-  PragmaNoTerminationCheck{}        -> PragmaNoTerminationCheck_
-  PragmaCompiled{}                  -> PragmaCompiled_
-  ShadowingInTelescope{}            -> ShadowingInTelescope_
-  UnknownFixityInMixfixDecl{}       -> UnknownFixityInMixfixDecl_
-  UnknownNamesInFixityDecl{}        -> UnknownNamesInFixityDecl_
-  UnknownNamesInPolarityPragmas{}   -> UnknownNamesInPolarityPragmas_
-  UselessAbstract{}                 -> UselessAbstract_
-  UselessInstance{}                 -> UselessInstance_
-  UselessPrivate{}                  -> UselessPrivate_
-
--- | Nicifier warnings turned into errors in @--safe@ mode.
-unsafeDeclarationWarning :: DeclarationWarning -> Bool
-unsafeDeclarationWarning = unsafeDeclarationWarning' . dwWarning
-
-unsafeDeclarationWarning' :: DeclarationWarning' -> Bool
-unsafeDeclarationWarning' = \case
-  -- Please keep in alphabetical order.
-  EmptyAbstract{}                   -> False
-  EmptyConstructor{}                -> False
-  EmptyField{}                      -> False
-  EmptyGeneralize{}                 -> False
-  EmptyInstance{}                   -> False
-  EmptyMacro{}                      -> False
-  EmptyMutual{}                     -> False
-  EmptyPrivate{}                    -> False
-  EmptyPostulate{}                  -> False
-  EmptyPrimitive{}                  -> False
-  InvalidCatchallPragma{}           -> False
-  InvalidConstructor{}              -> False
-  InvalidConstructorBlock{}         -> False
-  InvalidNoPositivityCheckPragma{}  -> False
-  InvalidNoUniverseCheckPragma{}    -> False
-  InvalidRecordDirective{}          -> False
-  InvalidTerminationCheckPragma{}   -> False
-  InvalidCoverageCheckPragma{}      -> False
-  MissingDefinitions{}              -> True  -- not safe
-  NotAllowedInMutual{}              -> False -- really safe?
-  OpenPublicPrivate{}               -> False
-  OpenPublicAbstract{}              -> False
-  PolarityPragmasButNotPostulates{} -> False
-  PragmaNoTerminationCheck{}        -> True  -- not safe
-  PragmaCompiled{}                  -> True  -- not safe
-  ShadowingInTelescope{}            -> False
-  UnknownFixityInMixfixDecl{}       -> False
-  UnknownNamesInFixityDecl{}        -> False
-  UnknownNamesInPolarityPragmas{}   -> False
-  UselessAbstract{}                 -> False
-  UselessInstance{}                 -> False
-  UselessPrivate{}                  -> False
-
--- | Several declarations expect only type signatures as sub-declarations.  These are:
-data KindOfBlock
-  = PostulateBlock  -- ^ @postulate@
-  | PrimitiveBlock  -- ^ @primitive@.  Ensured by parser.
-  | InstanceBlock   -- ^ @instance@.  Actually, here all kinds of sub-declarations are allowed a priori.
-  | FieldBlock      -- ^ @field@.  Ensured by parser.
-  | DataBlock       -- ^ @data ... where@.  Here we got a bad error message for Agda-2.5 (Issue 1698).
-  deriving (Data, Eq, Ord, Show)
-
-
-instance HasRange DeclarationException where
-  getRange (DeclarationException _ err) = getRange err
-
-instance HasRange DeclarationException' where
-  getRange (MultipleEllipses d)                 = getRange d
-  getRange (InvalidName x)                      = getRange x
-  getRange (DuplicateDefinition x)              = getRange x
-  getRange (DuplicateAnonDeclaration r)         = r
-  getRange (MissingWithClauses x lhs)           = getRange lhs
-  getRange (WrongDefinition x k k')             = getRange x
-  getRange (AmbiguousFunClauses lhs xs)         = getRange lhs
-  getRange (DeclarationPanic _)                 = noRange
-  getRange (WrongContentBlock _ r)              = r
-  getRange (InvalidMeasureMutual r)             = r
-  getRange (UnquoteDefRequiresSignature x)      = getRange x
-  getRange (BadMacroDef d)                      = getRange d
-
-instance HasRange DeclarationWarning where
-  getRange (DeclarationWarning _ w) = getRange w
-
-instance HasRange DeclarationWarning' where
-  getRange (UnknownNamesInFixityDecl xs)        = getRange xs
-  getRange (UnknownFixityInMixfixDecl xs)       = getRange xs
-  getRange (UnknownNamesInPolarityPragmas xs)   = getRange xs
-  getRange (PolarityPragmasButNotPostulates xs) = getRange xs
-  getRange (MissingDefinitions xs)              = getRange xs
-  getRange (UselessPrivate r)                   = r
-  getRange (NotAllowedInMutual r x)             = r
-  getRange (UselessAbstract r)                  = r
-  getRange (UselessInstance r)                  = r
-  getRange (EmptyConstructor r)                 = r
-  getRange (EmptyMutual r)                      = r
-  getRange (EmptyAbstract r)                    = r
-  getRange (EmptyPrivate r)                     = r
-  getRange (EmptyInstance r)                    = r
-  getRange (EmptyMacro r)                       = r
-  getRange (EmptyPostulate r)                   = r
-  getRange (EmptyGeneralize r)                  = r
-  getRange (EmptyPrimitive r)                   = r
-  getRange (EmptyField r)                       = r
-  getRange (InvalidTerminationCheckPragma r)    = r
-  getRange (InvalidCoverageCheckPragma r)       = r
-  getRange (InvalidNoPositivityCheckPragma r)   = r
-  getRange (InvalidCatchallPragma r)            = r
-  getRange (InvalidConstructor r)               = r
-  getRange (InvalidConstructorBlock r)          = r
-  getRange (InvalidNoUniverseCheckPragma r)     = r
-  getRange (InvalidRecordDirective r)           = r
-  getRange (PragmaNoTerminationCheck r)         = r
-  getRange (PragmaCompiled r)                   = r
-  getRange (OpenPublicAbstract r)               = r
-  getRange (OpenPublicPrivate r)                = r
-  getRange (ShadowingInTelescope ns)            = getRange ns
-
-instance HasRange NiceDeclaration where
-  getRange (Axiom r _ _ _ _ _ _)           = r
-  getRange (NiceField r _ _ _ _ _ _)       = r
-  getRange (NiceMutual r _ _ _ _)          = r
-  getRange (NiceModule r _ _ _ _ _ )       = r
-  getRange (NiceModuleMacro r _ _ _ _ _)   = r
-  getRange (NiceOpen r _ _)                = r
-  getRange (NiceImport r _ _ _ _)          = r
-  getRange (NicePragma r _)                = r
-  getRange (PrimitiveFunction r _ _ _ _)   = r
-  getRange (FunSig r _ _ _ _ _ _ _ _ _)    = r
-  getRange (FunDef r _ _ _ _ _ _ _)        = r
-  getRange (NiceDataDef r _ _ _ _ _ _ _)   = r
-  getRange (NiceLoneConstructor r _)       = r
-  getRange (NiceRecDef r _ _ _ _ _ _ _ _)  = r
-  getRange (NiceRecSig r _ _ _ _ _ _ _)    = r
-  getRange (NiceDataSig r _ _ _ _ _ _ _)   = r
-  getRange (NicePatternSyn r _ _ _ _)      = r
-  getRange (NiceGeneralize r _ _ _ _ _)    = r
-  getRange (NiceFunClause r _ _ _ _ _ _)   = r
-  getRange (NiceUnquoteDecl r _ _ _ _ _ _ _) = r
-  getRange (NiceUnquoteDef r _ _ _ _ _ _)    = r
-
-instance Pretty NiceDeclaration where
-  pretty = \case
-    Axiom _ _ _ _ _ x _            -> text "postulate" <+> pretty x <+> colon <+> text "_"
-    NiceField _ _ _ _ _ x _        -> text "field" <+> pretty x
-    PrimitiveFunction _ _ _ x _    -> text "primitive" <+> pretty x
-    NiceMutual{}                   -> text "mutual"
-    NiceModule _ _ _ x _ _         -> text "module" <+> pretty x <+> text "where"
-    NiceModuleMacro _ _ x _ _ _    -> text "module" <+> pretty x <+> text "= ..."
-    NiceOpen _ x _                 -> text "open" <+> pretty x
-    NiceImport _ x _ _ _           -> text "import" <+> pretty x
-    NicePragma{}                   -> text "{-# ... #-}"
-    NiceRecSig _ _ _ _ _ x _ _     -> text "record" <+> pretty x
-    NiceDataSig _ _ _ _ _ x _ _    -> text "data" <+> pretty x
-    NiceFunClause{}                -> text "<function clause>"
-    FunSig _ _ _ _ _ _ _ _ x _     -> pretty x <+> colon <+> text "_"
-    FunDef _ _ _ _ _ _ x _         -> pretty x <+> text "= _"
-    NiceDataDef _ _ _ _ _ x _ _    -> text "data" <+> pretty x <+> text "where"
-    NiceLoneConstructor _ ds       -> text "constructor"
-    NiceRecDef _ _ _ _ _ x  _ _ _  -> text "record" <+> pretty x <+> text "where"
-    NicePatternSyn _ _ x _ _       -> text "pattern" <+> pretty x
-    NiceGeneralize _ _ _ _ x _     -> text "variable" <+> pretty x
-    NiceUnquoteDecl _ _ _ _ _ _ xs _ -> text "<unquote declarations>"
-    NiceUnquoteDef _ _ _ _ _ xs _    -> text "<unquote definitions>"
-
--- These error messages can (should) be terminated by a dot ".",
--- there is no error context printed after them.
-instance Pretty DeclarationException' where
-  pretty (MultipleEllipses p) = fsep $
-    pwords "Multiple ellipses in left-hand side" ++ [pretty p]
-  pretty (InvalidName x) = fsep $
-    pwords "Invalid name:" ++ [pretty x]
-  pretty (DuplicateDefinition x) = fsep $
-    pwords "Duplicate definition of" ++ [pretty x]
-  pretty (DuplicateAnonDeclaration _) = fsep $
-    pwords "Duplicate declaration of _"
-  pretty (MissingWithClauses x lhs) = fsep $
-    pwords "Missing with-clauses for function" ++ [pretty x]
-
-  pretty (WrongDefinition x k k') = fsep $ pretty x :
-    pwords ("has been declared as a " ++ prettyShow k ++
-      ", but is being defined as a " ++ prettyShow k')
-  pretty (AmbiguousFunClauses lhs xs) = sep
-    [ fsep $
-        pwords "More than one matching type signature for left hand side " ++ [pretty lhs] ++
-        pwords "it could belong to any of:"
-    , vcat $ fmap (pretty . PrintRange) xs
-    ]
-  pretty (WrongContentBlock b _)      = fsep . pwords $
-    case b of
-      PostulateBlock -> "A postulate block can only contain type signatures, possibly under keyword instance"
-      DataBlock -> "A data definition can only contain type signatures, possibly under keyword instance"
-      _ -> "Unexpected declaration"
-  pretty (InvalidMeasureMutual _) = fsep $
-    pwords "In a mutual block, either all functions must have the same (or no) termination checking pragma."
-  pretty (UnquoteDefRequiresSignature xs) = fsep $
-    pwords "Missing type signatures for unquoteDef" ++ map pretty (List1.toList xs)
-  pretty (BadMacroDef nd) = fsep $
-    text (declName nd) : pwords "are not allowed in macro blocks"
-  pretty (DeclarationPanic s) = text s
-
-instance Pretty DeclarationWarning where
-  pretty (DeclarationWarning _ w) = pretty w
-
-instance Pretty DeclarationWarning' where
-  pretty (UnknownNamesInFixityDecl xs) = fsep $
-    pwords "The following names are not declared in the same scope as their syntax or fixity declaration (i.e., either not in scope at all, imported from another module, or declared in a super module):"
-    ++ punctuate comma (map pretty xs)
-  pretty (UnknownFixityInMixfixDecl xs) = fsep $
-    pwords "The following mixfix names do not have an associated fixity declaration:"
-    ++ punctuate comma (map pretty xs)
-  pretty (UnknownNamesInPolarityPragmas xs) = fsep $
-    pwords "The following names are not declared in the same scope as their polarity pragmas (they could for instance be out of scope, imported from another module, or declared in a super module):"
-    ++ punctuate comma  (map pretty xs)
-  pretty (MissingDefinitions xs) = fsep $
-   pwords "The following names are declared but not accompanied by a definition:"
-   ++ punctuate comma (map (pretty . fst) xs)
-  pretty (NotAllowedInMutual r nd) = fsep $
-    text nd : pwords "in mutual blocks are not supported.  Suggestion: get rid of the mutual block by manually ordering declarations"
-  pretty (PolarityPragmasButNotPostulates xs) = fsep $
-    pwords "Polarity pragmas have been given for the following identifiers which are not postulates:"
-    ++ punctuate comma (map pretty xs)
-  pretty (UselessPrivate _)      = fsep $
-    pwords "Using private here has no effect. Private applies only to declarations that introduce new identifiers into the module, like type signatures and data, record, and module declarations."
-  pretty (UselessAbstract _)      = fsep $
-    pwords "Using abstract here has no effect. Abstract applies to only definitions like data definitions, record type definitions and function clauses."
-  pretty (UselessInstance _)      = fsep $
-    pwords "Using instance here has no effect. Instance applies only to declarations that introduce new identifiers into the module, like type signatures and axioms."
-  pretty (EmptyMutual    _)  = fsep $ pwords "Empty mutual block."
-  pretty EmptyConstructor{}  = fsep $ pwords "Empty constructor block."
-  pretty (EmptyAbstract  _)  = fsep $ pwords "Empty abstract block."
-  pretty (EmptyPrivate   _)  = fsep $ pwords "Empty private block."
-  pretty (EmptyInstance  _)  = fsep $ pwords "Empty instance block."
-  pretty (EmptyMacro     _)  = fsep $ pwords "Empty macro block."
-  pretty (EmptyPostulate _)  = fsep $ pwords "Empty postulate block."
-  pretty (EmptyGeneralize _) = fsep $ pwords "Empty variable block."
-  pretty (EmptyPrimitive _)  = fsep $ pwords "Empty primitive block."
-  pretty (EmptyField _)      = fsep $ pwords "Empty field block."
-  pretty InvalidRecordDirective{} = fsep $
-    pwords "Record directives can only be used inside record definitions and before field declarations."
-  pretty (InvalidTerminationCheckPragma _) = fsep $
-    pwords "Termination checking pragmas can only precede a function definition or a mutual block (that contains a function definition)."
-  pretty InvalidConstructor{} = fsep $
-    pwords "`constructor' blocks may only contain type signatures for constructors."
-  pretty InvalidConstructorBlock{} = fsep $
-    pwords "No `constructor' blocks outside of `infix mutual' blocks."
-  pretty (InvalidCoverageCheckPragma _)    = fsep $
-    pwords "Coverage checking pragmas can only precede a function definition or a mutual block (that contains a function definition)."
-  pretty (InvalidNoPositivityCheckPragma _) = fsep $
-    pwords "NO_POSITIVITY_CHECKING pragmas can only precede a data/record definition or a mutual block (that contains a data/record definition)."
-  pretty (InvalidCatchallPragma _) = fsep $
-    pwords "The CATCHALL pragma can only precede a function clause."
-  pretty (InvalidNoUniverseCheckPragma _) = fsep $
-    pwords "NO_UNIVERSE_CHECKING pragmas can only precede a data/record definition."
-  pretty (PragmaNoTerminationCheck _) = fsep $
-    pwords "Pragma {-# NO_TERMINATION_CHECK #-} has been removed.  To skip the termination check, label your definitions either as {-# TERMINATING #-} or {-# NON_TERMINATING #-}."
-  pretty (PragmaCompiled _) = fsep $
-    pwords "COMPILE pragma not allowed in safe mode."
-  pretty (OpenPublicAbstract _) = fsep $
-    pwords "public does not have any effect in an abstract block."
-  pretty (OpenPublicPrivate _) = fsep $
-    pwords "public does not have any effect in a private block."
-  pretty (ShadowingInTelescope nrs) = fsep $
-    pwords "Shadowing in telescope, repeated variable names:"
-    ++ punctuate comma (map (pretty . fst) nrs)
-
-declName :: NiceDeclaration -> String
-declName Axiom{}             = "Postulates"
-declName NiceField{}         = "Fields"
-declName NiceMutual{}        = "Mutual blocks"
-declName NiceModule{}        = "Modules"
-declName NiceModuleMacro{}   = "Modules"
-declName NiceOpen{}          = "Open declarations"
-declName NiceImport{}        = "Import statements"
-declName NicePragma{}        = "Pragmas"
-declName PrimitiveFunction{} = "Primitive declarations"
-declName NicePatternSyn{}    = "Pattern synonyms"
-declName NiceGeneralize{}    = "Generalized variables"
-declName NiceUnquoteDecl{}   = "Unquoted declarations"
-declName NiceUnquoteDef{}    = "Unquoted definitions"
-declName NiceRecSig{}        = "Records"
-declName NiceDataSig{}       = "Data types"
-declName NiceFunClause{}     = "Functions without a type signature"
-declName FunSig{}            = "Type signatures"
-declName FunDef{}            = "Function definitions"
-declName NiceRecDef{}        = "Records"
-declName NiceDataDef{}       = "Data types"
-declName NiceLoneConstructor{} = "Constructors"
-
-{--------------------------------------------------------------------------
     The niceifier
  --------------------------------------------------------------------------}
-
-data InMutual
-  = InMutual    -- ^ we are nicifying a mutual block
-  | NotInMutual -- ^ we are nicifying decls not in a mutual block
-    deriving (Eq, Show)
-
--- | The kind of the forward declaration.
-
-data DataRecOrFun
-  = DataName
-    { _kindPosCheck :: PositivityCheck
-    , _kindUniCheck :: UniverseCheck
-    }
-    -- ^ Name of a data type
-  | RecName
-    { _kindPosCheck :: PositivityCheck
-    , _kindUniCheck :: UniverseCheck
-    }
-    -- ^ Name of a record type
-  | FunName TerminationCheck CoverageCheck
-    -- ^ Name of a function.
-  deriving (Data, Show)
-
--- Ignore pragmas when checking equality
-instance Eq DataRecOrFun where
-  DataName{} == DataName{} = True
-  RecName{}  == RecName{}  = True
-  FunName{}  == FunName{}  = True
-  _          == _          = False
-
-instance Pretty DataRecOrFun where
-  pretty DataName{} = "data type"
-  pretty RecName{}  = "record type"
-  pretty FunName{}  = "function"
-
-isFunName :: DataRecOrFun -> Bool
-isFunName (FunName{}) = True
-isFunName _           = False
-
-sameKind :: DataRecOrFun -> DataRecOrFun -> Bool
-sameKind = (==)
-
-terminationCheck :: DataRecOrFun -> TerminationCheck
-terminationCheck (FunName tc _) = tc
-terminationCheck _ = TerminationCheck
-
-coverageCheck :: DataRecOrFun -> CoverageCheck
-coverageCheck (FunName _ cc) = cc
-coverageCheck _ = YesCoverageCheck
-
-positivityCheck :: DataRecOrFun -> PositivityCheck
-positivityCheck (DataName pc _) = pc
-positivityCheck (RecName pc _)  = pc
-positivityCheck (FunName _ _)   = YesPositivityCheck
-
-universeCheck :: DataRecOrFun -> UniverseCheck
-universeCheck (DataName _ uc) = uc
-universeCheck (RecName _ uc)  = uc
-universeCheck (FunName _ _)   = YesUniverseCheck
 
 -- | Check that declarations in a mutual block are consistently
 --   equipped with MEASURE pragmas, or whether there is a
@@ -677,228 +141,6 @@ combineCoverageChecks = Fold.fold
 
 combinePositivityChecks :: [PositivityCheck] -> PositivityCheck
 combinePositivityChecks = Fold.fold
-
--- | Nicifier monad.
---   Preserve the state when throwing an exception.
-
-newtype Nice a = Nice { unNice :: ExceptT DeclarationException (State NiceEnv) a }
-  deriving ( Functor, Applicative, Monad
-           , MonadState NiceEnv, MonadError DeclarationException
-           )
-
--- | Run a Nicifier computation, return result and warnings
---   (in chronological order).
-runNice :: Nice a -> (Either DeclarationException a, NiceWarnings)
-runNice m = second (reverse . niceWarn) $
-  runExceptT (unNice m) `runState` initNiceEnv
-
--- | Nicifier state.
-
-data NiceEnv = NiceEnv
-  { _loneSigs :: LoneSigs
-    -- ^ Lone type signatures that wait for their definition.
-  , _termChk  :: TerminationCheck
-    -- ^ Termination checking pragma waiting for a definition.
-  , _posChk   :: PositivityCheck
-    -- ^ Positivity checking pragma waiting for a definition.
-  , _uniChk   :: UniverseCheck
-    -- ^ Universe checking pragma waiting for a data/rec signature or definition.
-  , _catchall :: Catchall
-    -- ^ Catchall pragma waiting for a function clause.
-  , _covChk  :: CoverageCheck
-    -- ^ Coverage pragma waiting for a definition.
-  , niceWarn :: NiceWarnings
-    -- ^ Stack of warnings. Head is last warning.
-  , _nameId  :: NameId
-    -- ^ We distinguish different 'NoName's (anonymous definitions) by a unique 'NameId'.
-  }
-
-data LoneSig = LoneSig
-  { loneSigRange :: Range
-  , loneSigName  :: Name
-      -- ^ If 'isNoName', this name can have a different 'NameId'
-      --   than the key of 'LoneSigs' pointing to it.
-  , loneSigKind  :: DataRecOrFun
-  }
-
-type LoneSigs     = Map Name LoneSig
-     -- ^ We retain the 'Name' also in the codomain since
-     --   'Name' as a key is up to @Eq Name@ which ignores the range.
-     --   However, without range names are not unique in case the
-     --   user gives a second definition of the same name.
-     --   This causes then problems in 'replaceSigs' which might
-     --   replace the wrong signature.
-     --
-     --   Another reason is that we want to distinguish different
-     --   occurrences of 'NoName' in a mutual block (issue #4157).
-     --   The 'NoName' in the codomain will have a unique 'NameId'.
-
-type NiceWarnings = [DeclarationWarning]
-     -- ^ Stack of warnings. Head is last warning.
-
--- | Initial nicifier state.
-
-initNiceEnv :: NiceEnv
-initNiceEnv = NiceEnv
-  { _loneSigs = empty
-  , _termChk  = TerminationCheck
-  , _posChk   = YesPositivityCheck
-  , _uniChk   = YesUniverseCheck
-  , _catchall = False
-  , _covChk   = YesCoverageCheck
-  , niceWarn  = []
-  , _nameId   = NameId 1 1   -- The module id is bogus.
-  }
-
-lensNameId :: Lens' NameId NiceEnv
-lensNameId f e = f (_nameId e) <&> \ i -> e { _nameId = i }
-
-nextNameId :: Nice NameId
-nextNameId = do
-  i <- use lensNameId
-  lensNameId %= succ
-  return i
-
--- * Handling the lone signatures, stored to infer mutual blocks.
-
--- | Lens for field '_loneSigs'.
-
-loneSigs :: Lens' LoneSigs NiceEnv
-loneSigs f e = f (_loneSigs e) <&> \ s -> e { _loneSigs = s }
-
--- | Adding a lone signature to the state.
---   Return the name (which is made unique if 'isNoName').
-
-addLoneSig :: Range -> Name -> DataRecOrFun -> Nice Name
-addLoneSig r x k = do
-  -- Andreas, 2020-05-19, issue #4157, make '_' unique.
-  x' <- if not $ isNoName x then return x else do
-    i <- nextNameId
-    return x{ nameId = i }
-  loneSigs %== \ s -> do
-    let (mr, s') = Map.insertLookupWithKey (\ _k new _old -> new) x (LoneSig r x' k) s
-    case mr of
-      Nothing -> return s'
-      Just{}  -> declarationException $
-        if not $ isNoName x then DuplicateDefinition x else DuplicateAnonDeclaration r
-  return x'
-
--- | Remove a lone signature from the state.
-
-removeLoneSig :: Name -> Nice ()
-removeLoneSig x = loneSigs %= Map.delete x
-
--- | Search for forward type signature.
-
-getSig :: Name -> Nice (Maybe DataRecOrFun)
-getSig x = fmap loneSigKind . Map.lookup x <$> use loneSigs
-
--- | Check that no lone signatures are left in the state.
-
-noLoneSigs :: Nice Bool
-noLoneSigs = null <$> use loneSigs
-
--- | Ensure that all forward declarations have been given a definition.
-
-forgetLoneSigs :: Nice ()
-forgetLoneSigs = loneSigs .= Map.empty
-
-checkLoneSigs :: LoneSigs -> Nice ()
-checkLoneSigs xs = do
-  forgetLoneSigs
-  unless (Map.null xs) $ declarationWarning $ MissingDefinitions $
-    map (\s -> (loneSigName s , loneSigRange s)) $ Map.elems xs
-
--- | Get names of lone function signatures, plus their unique names.
-
-loneFuns :: LoneSigs -> [(Name,Name)]
-loneFuns = map (second loneSigName) . filter (isFunName . loneSigKind . snd) . Map.toList
-
--- | Create a 'LoneSigs' map from an association list.
-
-loneSigsFromLoneNames :: [(Range, Name, DataRecOrFun)] -> LoneSigs
-loneSigsFromLoneNames = Map.fromList . map (\(r,x,k) -> (x, LoneSig r x k))
-
--- | Lens for field '_termChk'.
-
-terminationCheckPragma :: Lens' TerminationCheck NiceEnv
-terminationCheckPragma f e = f (_termChk e) <&> \ s -> e { _termChk = s }
-
-withTerminationCheckPragma :: TerminationCheck -> Nice a -> Nice a
-withTerminationCheckPragma tc f = do
-  tc_old <- use terminationCheckPragma
-  terminationCheckPragma .= tc
-  result <- f
-  terminationCheckPragma .= tc_old
-  return result
-
-coverageCheckPragma :: Lens' CoverageCheck NiceEnv
-coverageCheckPragma f e = f (_covChk e) <&> \ s -> e { _covChk = s }
-
-withCoverageCheckPragma :: CoverageCheck -> Nice a -> Nice a
-withCoverageCheckPragma tc f = do
-  tc_old <- use coverageCheckPragma
-  coverageCheckPragma .= tc
-  result <- f
-  coverageCheckPragma .= tc_old
-  return result
-
--- | Lens for field '_posChk'.
-
-positivityCheckPragma :: Lens' PositivityCheck NiceEnv
-positivityCheckPragma f e = f (_posChk e) <&> \ s -> e { _posChk = s }
-
-withPositivityCheckPragma :: PositivityCheck -> Nice a -> Nice a
-withPositivityCheckPragma pc f = do
-  pc_old <- use positivityCheckPragma
-  positivityCheckPragma .= pc
-  result <- f
-  positivityCheckPragma .= pc_old
-  return result
-
--- | Lens for field '_uniChk'.
-
-universeCheckPragma :: Lens' UniverseCheck NiceEnv
-universeCheckPragma f e = f (_uniChk e) <&> \ s -> e { _uniChk = s }
-
-withUniverseCheckPragma :: UniverseCheck -> Nice a -> Nice a
-withUniverseCheckPragma uc f = do
-  uc_old <- use universeCheckPragma
-  universeCheckPragma .= uc
-  result <- f
-  universeCheckPragma .= uc_old
-  return result
-
--- | Get universe check pragma from a data/rec signature.
---   Defaults to 'YesUniverseCheck'.
-
-getUniverseCheckFromSig :: Name -> Nice UniverseCheck
-getUniverseCheckFromSig x = maybe YesUniverseCheck universeCheck <$> getSig x
-
--- | Lens for field '_catchall'.
-
-catchallPragma :: Lens' Catchall NiceEnv
-catchallPragma f e = f (_catchall e) <&> \ s -> e { _catchall = s }
-
--- | Get current catchall pragma, and reset it for the next clause.
-
-popCatchallPragma :: Nice Catchall
-popCatchallPragma = do
-  ca <- use catchallPragma
-  catchallPragma .= False
-  return ca
-
-withCatchallPragma :: Catchall -> Nice a -> Nice a
-withCatchallPragma ca f = do
-  ca_old <- use catchallPragma
-  catchallPragma .= ca
-  result <- f
-  catchallPragma .= ca_old
-  return result
-
--- | Add a new warning.
-niceWarning :: DeclarationWarning -> Nice ()
-niceWarning w = modify $ \ st -> st { niceWarn = w : niceWarn st }
 
 data DeclKind
     = LoneSigDecl Range DataRecOrFun Name
