+++ src/full/Agda/Syntax/Concrete/Definitions/Types.hs
@@ -1,0 +1,233 @@
+module Agda.Syntax.Concrete.Definitions.Types where
+
+import Data.Data
+
+import Agda.Syntax.Position
+import Agda.Syntax.Common hiding (TerminationCheck())
+import qualified Agda.Syntax.Common as Common
+import Agda.Syntax.Concrete
+import Agda.Syntax.Concrete.Name
+import Agda.Syntax.Concrete.Pretty
+
+import Agda.Utils.Pretty
+
+{--------------------------------------------------------------------------
+    Types
+ --------------------------------------------------------------------------}
+
+{-| The nice declarations. No fixity declarations and function definitions are
+    contained in a single constructor instead of spread out between type
+    signatures and clauses. The @private@, @postulate@, @abstract@ and @instance@
+    modifiers have been distributed to the individual declarations.
+
+    Observe the order of components:
+
+      Range
+      Fixity'
+      Access
+      IsAbstract
+      IsInstance
+      TerminationCheck
+      PositivityCheck
+
+      further attributes
+
+      (Q)Name
+
+      content (Expr, Declaration ...)
+-}
+data NiceDeclaration
+  = Axiom Range Access IsAbstract IsInstance ArgInfo Name Expr
+      -- ^ 'IsAbstract' argument: We record whether a declaration was made in an @abstract@ block.
+      --
+      --   'ArgInfo' argument: Axioms and functions can be declared irrelevant.
+      --   ('Hiding' should be 'NotHidden'.)
+  | NiceField Range Access IsAbstract IsInstance TacticAttribute Name (Arg Expr)
+  | PrimitiveFunction Range Access IsAbstract Name (Arg Expr)
+  | NiceMutual Range TerminationCheck CoverageCheck PositivityCheck [NiceDeclaration]
+  | NiceModule Range Access IsAbstract QName Telescope [Declaration]
+  | NiceModuleMacro Range Access Name ModuleApplication OpenShortHand ImportDirective
+  | NiceOpen Range QName ImportDirective
+  | NiceImport Range QName (Maybe AsName) OpenShortHand ImportDirective
+  | NicePragma Range Pragma
+  | NiceRecSig Range Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr
+  | NiceDataSig Range Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr
+  | NiceFunClause Range Access IsAbstract TerminationCheck CoverageCheck Catchall Declaration
+    -- ^ An uncategorized function clause, could be a function clause
+    --   without type signature or a pattern lhs (e.g. for irrefutable let).
+    --   The 'Declaration' is the actual 'FunClause'.
+  | FunSig Range Access IsAbstract IsInstance IsMacro ArgInfo TerminationCheck CoverageCheck Name Expr
+  | FunDef Range [Declaration] IsAbstract IsInstance TerminationCheck CoverageCheck Name [Clause]
+      -- ^ Block of function clauses (we have seen the type signature before).
+      --   The 'Declaration's are the original declarations that were processed
+      --   into this 'FunDef' and are only used in 'notSoNiceDeclaration'.
+      --   Andreas, 2017-01-01: Because of issue #2372, we add 'IsInstance' here.
+      --   An alias should know that it is an instance.
+  | NiceDataDef Range Origin IsAbstract PositivityCheck UniverseCheck Name [LamBinding] [NiceConstructor]
+  | NiceLoneConstructor Range [NiceConstructor]
+  | NiceRecDef Range Origin IsAbstract PositivityCheck UniverseCheck Name RecordDirectives [LamBinding] [Declaration]
+      -- ^ @(Maybe Range)@ gives range of the 'pattern' declaration.
+  | NicePatternSyn Range Access Name [Arg Name] Pattern
+  | NiceGeneralize Range Access ArgInfo TacticAttribute Name Expr
+  | NiceUnquoteDecl Range Access IsAbstract IsInstance TerminationCheck CoverageCheck [Name] Expr
+  | NiceUnquoteDef Range Access IsAbstract TerminationCheck CoverageCheck [Name] Expr
+  deriving (Data, Show)
+
+type TerminationCheck = Common.TerminationCheck Measure
+
+-- | Termination measure is, for now, a variable name.
+type Measure = Name
+
+type Catchall = Bool
+
+-- | Only 'Axiom's.
+type NiceConstructor = NiceTypeSignature
+
+-- | Only 'Axiom's.
+type NiceTypeSignature  = NiceDeclaration
+
+-- | One clause in a function definition. There is no guarantee that the 'LHS'
+--   actually declares the 'Name'. We will have to check that later.
+data Clause = Clause Name Catchall LHS RHS WhereClause [Clause]
+    deriving (Data, Show)
+
+
+-- | Several declarations expect only type signatures as sub-declarations.  These are:
+data KindOfBlock
+  = PostulateBlock  -- ^ @postulate@
+  | PrimitiveBlock  -- ^ @primitive@.  Ensured by parser.
+  | InstanceBlock   -- ^ @instance@.  Actually, here all kinds of sub-declarations are allowed a priori.
+  | FieldBlock      -- ^ @field@.  Ensured by parser.
+  | DataBlock       -- ^ @data ... where@.  Here we got a bad error message for Agda-2.5 (Issue 1698).
+  deriving (Data, Eq, Ord, Show)
+
+
+instance HasRange NiceDeclaration where
+  getRange (Axiom r _ _ _ _ _ _)           = r
+  getRange (NiceField r _ _ _ _ _ _)       = r
+  getRange (NiceMutual r _ _ _ _)          = r
+  getRange (NiceModule r _ _ _ _ _ )       = r
+  getRange (NiceModuleMacro r _ _ _ _ _)   = r
+  getRange (NiceOpen r _ _)                = r
+  getRange (NiceImport r _ _ _ _)          = r
+  getRange (NicePragma r _)                = r
+  getRange (PrimitiveFunction r _ _ _ _)   = r
+  getRange (FunSig r _ _ _ _ _ _ _ _ _)    = r
+  getRange (FunDef r _ _ _ _ _ _ _)        = r
+  getRange (NiceDataDef r _ _ _ _ _ _ _)   = r
+  getRange (NiceLoneConstructor r _)       = r
+  getRange (NiceRecDef r _ _ _ _ _ _ _ _)  = r
+  getRange (NiceRecSig r _ _ _ _ _ _ _)    = r
+  getRange (NiceDataSig r _ _ _ _ _ _ _)   = r
+  getRange (NicePatternSyn r _ _ _ _)      = r
+  getRange (NiceGeneralize r _ _ _ _ _)    = r
+  getRange (NiceFunClause r _ _ _ _ _ _)   = r
+  getRange (NiceUnquoteDecl r _ _ _ _ _ _ _) = r
+  getRange (NiceUnquoteDef r _ _ _ _ _ _)    = r
+
+instance Pretty NiceDeclaration where
+  pretty = \case
+    Axiom _ _ _ _ _ x _            -> text "postulate" <+> pretty x <+> colon <+> text "_"
+    NiceField _ _ _ _ _ x _        -> text "field" <+> pretty x
+    PrimitiveFunction _ _ _ x _    -> text "primitive" <+> pretty x
+    NiceMutual{}                   -> text "mutual"
+    NiceModule _ _ _ x _ _         -> text "module" <+> pretty x <+> text "where"
+    NiceModuleMacro _ _ x _ _ _    -> text "module" <+> pretty x <+> text "= ..."
+    NiceOpen _ x _                 -> text "open" <+> pretty x
+    NiceImport _ x _ _ _           -> text "import" <+> pretty x
+    NicePragma{}                   -> text "{-# ... #-}"
+    NiceRecSig _ _ _ _ _ x _ _     -> text "record" <+> pretty x
+    NiceDataSig _ _ _ _ _ x _ _    -> text "data" <+> pretty x
+    NiceFunClause{}                -> text "<function clause>"
+    FunSig _ _ _ _ _ _ _ _ x _     -> pretty x <+> colon <+> text "_"
+    FunDef _ _ _ _ _ _ x _         -> pretty x <+> text "= _"
+    NiceDataDef _ _ _ _ _ x _ _    -> text "data" <+> pretty x <+> text "where"
+    NiceLoneConstructor _ ds       -> text "constructor"
+    NiceRecDef _ _ _ _ _ x  _ _ _  -> text "record" <+> pretty x <+> text "where"
+    NicePatternSyn _ _ x _ _       -> text "pattern" <+> pretty x
+    NiceGeneralize _ _ _ _ x _     -> text "variable" <+> pretty x
+    NiceUnquoteDecl _ _ _ _ _ _ xs _ -> text "<unquote declarations>"
+    NiceUnquoteDef _ _ _ _ _ xs _    -> text "<unquote definitions>"
+
+declName :: NiceDeclaration -> String
+declName Axiom{}             = "Postulates"
+declName NiceField{}         = "Fields"
+declName NiceMutual{}        = "Mutual blocks"
+declName NiceModule{}        = "Modules"
+declName NiceModuleMacro{}   = "Modules"
+declName NiceOpen{}          = "Open declarations"
+declName NiceImport{}        = "Import statements"
+declName NicePragma{}        = "Pragmas"
+declName PrimitiveFunction{} = "Primitive declarations"
+declName NicePatternSyn{}    = "Pattern synonyms"
+declName NiceGeneralize{}    = "Generalized variables"
+declName NiceUnquoteDecl{}   = "Unquoted declarations"
+declName NiceUnquoteDef{}    = "Unquoted definitions"
+declName NiceRecSig{}        = "Records"
+declName NiceDataSig{}       = "Data types"
+declName NiceFunClause{}     = "Functions without a type signature"
+declName FunSig{}            = "Type signatures"
+declName FunDef{}            = "Function definitions"
+declName NiceRecDef{}        = "Records"
+declName NiceDataDef{}       = "Data types"
+declName NiceLoneConstructor{} = "Constructors"
+
+
+data InMutual
+  = InMutual    -- ^ we are nicifying a mutual block
+  | NotInMutual -- ^ we are nicifying decls not in a mutual block
+    deriving (Eq, Show)
+
+-- | The kind of the forward declaration.
+
+data DataRecOrFun
+  = DataName
+    { _kindPosCheck :: PositivityCheck
+    , _kindUniCheck :: UniverseCheck
+    }
+    -- ^ Name of a data type
+  | RecName
+    { _kindPosCheck :: PositivityCheck
+    , _kindUniCheck :: UniverseCheck
+    }
+    -- ^ Name of a record type
+  | FunName TerminationCheck CoverageCheck
+    -- ^ Name of a function.
+  deriving (Data, Show)
+
+-- Ignore pragmas when checking equality
+instance Eq DataRecOrFun where
+  DataName{} == DataName{} = True
+  RecName{}  == RecName{}  = True
+  FunName{}  == FunName{}  = True
+  _          == _          = False
+
+instance Pretty DataRecOrFun where
+  pretty DataName{} = "data type"
+  pretty RecName{}  = "record type"
+  pretty FunName{}  = "function"
+
+isFunName :: DataRecOrFun -> Bool
+isFunName (FunName{}) = True
+isFunName _           = False
+
+sameKind :: DataRecOrFun -> DataRecOrFun -> Bool
+sameKind = (==)
+
+terminationCheck :: DataRecOrFun -> TerminationCheck
+terminationCheck (FunName tc _) = tc
+terminationCheck _ = TerminationCheck
+
+coverageCheck :: DataRecOrFun -> CoverageCheck
+coverageCheck (FunName _ cc) = cc
+coverageCheck _ = YesCoverageCheck
+
+positivityCheck :: DataRecOrFun -> PositivityCheck
+positivityCheck (DataName pc _) = pc
+positivityCheck (RecName pc _)  = pc
+positivityCheck (FunName _ _)   = YesPositivityCheck
+
+universeCheck :: DataRecOrFun -> UniverseCheck
+universeCheck (DataName _ uc) = uc
+universeCheck (RecName _ uc)  = uc
+universeCheck (FunName _ _)   = YesUniverseCheck
